Add files via upload
This is a basic python program that can convert your phrase into a sentence. If you have added a question in form of phrase the program capitalizes the first letter of the sentence and adds a question mark at the end. If it is a normal sentence it adds a full stop at the end. Using the /stop command ends the process and you receive all your sentences together with mentioned changes. So something like this:
Enter your msg: how are you
Enter your msg: i am fine
Enter your msg: /stop
Output: How are you? I am fine.